login & auth update - going to move to new logreg branch for next login, register and auth changes
