+++ src/styles/Components/navBar.scss
@@ -1,0 +1,8 @@
+.navbar{
+  // box-shadow: 0px 2px 10px rgba(0,0,0,0.2);
+  background-color: wheat;
+  .nav-link, span{
+    margin-left: 10px;
+    cursor: pointer;
+  }
+}
