+++ src/styles/main.scss
@@ -1,5 +1,6 @@
 @import './base';
 @import './variables';
 @import './components/movieInfo';
+@import './components/navBar.scss';
 @import './components/stickyFooter.scss';
